Initial Commit of Sphinx Config project
Creates a lfdocs-conf package that is pip installable.

The minimal conf.yaml consists of only a project_cfg configuration
with the name of the project eg opendaylight, opnfv, lfdocs:

   ---
   project_cfg: lfdocs

Limitations:

intersphinx_mapping is not currently mappable via yaml due to
it being a complex data set of a tuple + data type (None). Projects
can override this by setting it in their local conf.py or accept the
defaults provided by the lfdocs-conf project.

Change-Id: Ia695047dc92adc00e4e9ec1f6d3ed824d10c0959
Co-authored-by: Thanh Ha <thanh.ha@linuxfoundation.org>
Signed-off-by: Trevor Bramwell <tbramwell@linuxfoundation.org>
Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>