Move code out of __init__ and into conf.py module
This allows us to use __init__.py to store __version__ and other
metadata strings and be reused in local documentation and setup.py
to shore a single version file.

The problem with having code in __init__.py is that the code will
be executed even if all the documentation code wants to pull is the
metadata variable.

Change-Id: I7382c81508be42a41c99f392f8006cfdf9887258
Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>