Conditionally import OPNFV Theme
When using the 'opnfv' theme, conf.py needs to run the Python code
'sphinx_opnfv_theme.get_html_theme_path()' in order
to determine the location of package data on the filesystem.

Instead of having every project that wants to use the OPNFV theme set
this explicitly in their conf.py, we conditionally import the theme and
execute the code.

Change-Id: Ic686483fd511bf5e321977c2802f4cfb8986c800
Signed-off-by: Trevor Bramwell <tbramwell@linuxfoundation.org>