Testing Framework
This is currently broken due to imports. The yaml file get rewritten
before being imported (not a new tmp dir) which does nothing as Python
won't reimport the module.

Not sure yet if expliclity creating a tempdirectory for each test will
fix this (though be slow), or if we need to find a way to 'unload' the
module between tests.

Change-Id: I6f0722b3fded3f2751d4a2286a75afe7ae7e3cfc
Signed-off-by: Trevor Bramwell <tbramwell@linuxfoundation.org>