Pin Sphinx to version <1.8.0
The new 1.8.0 release of Sphinx is breaking builds in our CI.

In order to get tox runs to pass, the requirement files need to be split
apart and Sphinx version explicitly set for coala. This is due to some
coala bear requiring a version of Sphinx <1.5, which appears to get
pulled in from the base requirements.txt file.

This patch also includes minor clean up of the tox.ini file and a
pyflake validation fix.

Change-Id: Iea9b870270d7a8954f611aaca451de653397397a
Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>
Signed-off-by: Trevor Bramwell <tbramwell@linuxfoundation.org>