Add black to pre-commit
* Requires Python 3.6
* Changes pep8 max line length to 88 to match black
* Add .editorconfig to match black config
* Update tox to 2.0 to match black recommended version

This patch uses "black" to automatically manage Python style
conventions so that we don't have to. It utilizes pre-commit
to automatically handle running black on commit so that a
contributor does not have to think about it.

This however means that there is a required install step to
initialize a new dev environment using one of 2 options below:

1) Install pre-commit via your pkg manager then run:

    pre-commit install

2) Use tox to install pre-commit:

    tox -e check-hooks

Change-Id: I527e58b763dc7b0fdfa60fa619c61d2c63356d47
Signed-off-by: Thanh Ha <zxiiro@gmail.com>