Replace coala commit-msg hook with pre-commit
Idea here is to replace the slow coala checker with the
much faster pre-commit hooks in the future as we replace
the different linters we're using via coala with pre-commit
versions.

Using this requires the developer to setup pre-commit on their
local dev environment to initialize this the first time the
contribute to this repo.

Install can be performed via one of the following methods:

1) Install pre-commit from your local package manager then run:

    pre-commit install --hook-type commit-msg

2) Run `tox -e check-hooks`

Change-Id: I5b5d3dcbd7d83e90322f63ba99fa1f76c88ee8ad
Signed-off-by: Thanh Ha <zxiiro@gmail.com>