Migrate yamllint to pre-commit hook
We'd like to move off of Coala as it hasn't had a release in a very
long time. This patch sets up the pre-commit configuration for the
yamllint hook as well as "prettier" for automated file formatting.

This patch also sets YAML indents to 2 as it is the corrent dominant
indent size for the YAML files in this project.

Issue: RELENG-2642
Change-Id: I3827d922023ac18519d94c906b211274ff3e65d9
Signed-off-by: Thanh Ha <zxiiro@gmail.com>