CI: Use gitlint-ci at manual stage in pre-commit
When running pre-commit in a CI system the COMMIT_EDITMSG does not
(normally) get created as that is an artifact of editing the commit
message. If the file does not exist then gitlint will skip which makes
it possible for pre-commit checks that should fail, to pass.

The initial workaround proposed in global-jjb was to generate such a
file but it had several drawbacks such as:
- the necessity to pass the HOME variable to virtualenv even locally
- /bin/sh as whitelist_externals
- conditional scripting that might have unexpected side-effects in
  another situation
  (e.g. a git configuration initiated from an anonymous git clone)
https://gerrit.linuxfoundation.org/infra/c/releng/global-jjb/+/67641

The pre-commit hook provided up to now has a different behavior
than standalone gitlint defauts settings that are advised in a CI
context (i.e. "gitlint lint" or simply "gitlint").
https://jorisroovers.com/gitlint/#using-gitlint-in-a-ci-environment

The parameters --stage and --msg-filename options are passed and they
appear to be the root cause of the problems in the CI since they add
new requirements i.e. a local git user configuration and the presence of
an extra file argument.

Indeed, the issue should better be fixed in gitlint uptream repository
itself so that gitlint can be run at the commit or manual stage of
pre-commit.
This is now the case with a new 'gilint-ci' profile currently available
in gitlint official master branch only.
It will be released with the next version tagged 0.18.0
https://github.com/jorisroovers/gitlint/pull/192

- update gitlint revision in pre-commit configuration
- declare the new id "gitlint-ci" in pre-commit configuration
  at the manual stage
- adapt tox configuration accordingly

Change-Id: If9c99bfff02dfe4d118d0e1d8d35e9e646726170
Signed-off-by: Guillaume Lambert <guillaume.lambert@orange.com>