CI: Use gitlint at commit stage from pre-commit
- fork Joris Roover's Gitlint official repository on github
- push into it a new commit that adds a new id "gitlint0" in pre-commit
  hook configuration and allows to run gitlint at the commit stage.
- reference here this new commit and the forked repository
- configure the new id "gitlint0" to be run at the commit stage
- adapt tox configuration accordingly

When running pre-commit in a CI system the COMMIT_EDITMSG does not
(normally) get created as that is an artifact of editing the commit
message. If the file does not exist then gitlint will skip which makes
it possible for pre-commit checks that should fail, to pass.

The initial workaround proposed in global-jjb was to generate such a
file but it had several drawbacks such as:
- the necessity to pass the HOME variable to virtualenv even locally
- /bin/sh as whitelist_externals
- conditional scripting that might have unexpected side-effects in
  another situation.
https://gerrit.linuxfoundation.org/infra/c/releng/global-jjb/+/67641

Indeed, the issue should better be fixed in gitlint uptream repository
itself so that gitlint can be run at the commit stage from pre-commit.
Since gitlint embeds its own hook solution for this stage, this
configuration is not proposed in gitlint official repository, very
likely because this need was never felt.
https://jorisroovers.com/gitlint/#using-gitlint-through-pre-commit

Also the pre-commit hook provided up to now has a different behavior
than standalone gitlint defauts settings that are advised in a CI
context(i.e. "gitlint lint" or simply "gitlint").
https://jorisroovers.com/gitlint/#using-gitlint-in-a-ci-environment

The parameters --stage and --msg-filename options are passed and they
appears to be the root cause of the problems in the CI since they add
new requirements i.e. a local git user configuration and the presence of
an extra file argument.

https://github.com/guillaumelambert/gitlint/blob/main/.pre-commit-hooks.yaml
introduced in commit 9de1f89b8697809713bce0eadf25fdd578b1d844

pre-commit hook YAML upstream configuration does not allow to use
different options depending on the stage inside a same id block nor with
two blocks using the same id.
Only options from the last block appears to be used in that case.
As a consequence, a new id "gitlint0" must be declared but it still can
use the same name "gilint" displayed at runtime.

Change-Id: If9c99bfff02dfe4d118d0e1d8d35e9e646726170
Signed-off-by: Guillaume Lambert <guillaume.lambert@orange.com>